+++ src/test/testData/annotation_zuban/main.py
@@ -1,0 +1,2 @@
+from pkg.mod import VALUE
+print(VALUE)