+++ src/test/testData/annotation_zuban/pkg/mod.py
@@ -1,0 +1,1 @@
+VALUE: int = "oops"  # E_MARKER